Re-org StaticIp view
